test(search): split source revision integration fixture
